1.DashBoard 에서 복습현황보기 빼고 다 구현

+++ src/main/java/PlanQ/PlanQ/question/Question.java
@@ -37,6 +37,6 @@ public class Question {
     private Long selectOption;
 
     @Column(nullable = false)
-    private Long num;
+    private int num;
 
 }
